Fix model address, add screen statistics, just statistics by day of list post

+++ HuTiAdmin/Models/Post.swift
@@ -24,7 +24,7 @@ struct Post: Codable {
     var long: Double
     var userId: String
     var userName: String?
-    var createdAt: Double
+    var createdAt: Int
     var thumbnail: String? /// lấy image đầu tiên trong mảng images
     
     func getFullAddress() -> String {
